Move nfs plugin
Move the nfs plugin files to $(pythondir)/nfs_plugin
https://github.com/libstorage/libstoragemgmt/issues/440
