+++ plugin/nfs_plugin/__init__.py
@@ -1,0 +1,1 @@
+from nfs_plugin.nfs import NFSPlugin
